tests/submission/PhilipTaylor/canvas/2d.line.join.bevel.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.line.join.bevel</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.join.html">join</a>.bevel</h1>
<p class="desc">lineJoin 'bevel' is rendered correctly</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.lineJoin.common">2d.lineJoin.common</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.2d.lineJoin.bevel">2d.lineJoin.bevel</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("lineJoin 'bevel' is rendered correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

var tol = 1; // tolerance to avoid antialiasing artifacts

ctx.lineJoin = 'bevel';
ctx.lineWidth = 20;

ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';

ctx.fillRect(10, 10, 20, 20);
ctx.fillRect(20, 20, 20, 20);
ctx.beginPath();
ctx.moveTo(30, 20);
ctx.lineTo(40-tol, 20);
ctx.lineTo(30, 10+tol);
ctx.fill();

ctx.beginPath();
ctx.moveTo(10, 20);
ctx.lineTo(30, 20);
ctx.lineTo(30, 40);
ctx.stroke();


ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';

ctx.beginPath();
ctx.moveTo(60, 20);
ctx.lineTo(80, 20);
ctx.lineTo(80, 40);
ctx.stroke();

ctx.fillRect(60, 10, 20, 20);
ctx.fillRect(70, 20, 20, 20);
ctx.beginPath();
ctx.moveTo(80, 20);
ctx.lineTo(90+tol, 20);
ctx.lineTo(80, 10-tol);
ctx.fill();

_assertPixel(canvas, 34,16, 0,255,0,255, "34,16", "0,255,0,255");
_assertPixel(canvas, 34,15, 0,255,0,255, "34,15", "0,255,0,255");
_assertPixel(canvas, 35,15, 0,255,0,255, "35,15", "0,255,0,255");
_assertPixel(canvas, 36,15, 0,255,0,255, "36,15", "0,255,0,255");
_assertPixel(canvas, 36,14, 0,255,0,255, "36,14", "0,255,0,255");

_assertPixel(canvas, 84,16, 0,255,0,255, "84,16", "0,255,0,255");
_assertPixel(canvas, 84,15, 0,255,0,255, "84,15", "0,255,0,255");
_assertPixel(canvas, 85,15, 0,255,0,255, "85,15", "0,255,0,255");
_assertPixel(canvas, 86,15, 0,255,0,255, "86,15", "0,255,0,255");
_assertPixel(canvas, 86,14, 0,255,0,255, "86,14", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
